+++ src/ios/ContactX.swift
@@ -3,23 +3,54 @@ import Contacts
 class ContactX {
 
     var contact: CNContact;
+    var options: ContactsXOptions;
 
-    init(contact: CNContact) {
+    init(contact: CNContact, options: ContactsXOptions) {
         self.contact = contact
+        self.options = options;
+    }
+
+    func getEmailAddresses() -> [NSDictionary] {
+        let labeledValues: [NSDictionary] = self.contact.emailAddresses.map { (ob: CNLabeledValue<NSString>) -> NSDictionary in
+            return [
+                "id": ob.identifier,
+                "type": CNLabeledValue<NSString>.localizedString(forLabel: ob.label ?? ""),
+                "value": ob.value
+            ]
+        }
+        return labeledValues;
     }
 
     func getJson() -> NSDictionary {
 
-        let phoneNumbers: [String] = self.contact.phoneNumbers.map { (ob: CNLabeledValue<CNPhoneNumber>) -> String in
-            return ob.value.stringValue
+        var phoneNumbers: [String] = [];
+        if(options.phoneNumbers) {
+            phoneNumbers = self.contact.phoneNumbers.map { (ob: CNLabeledValue<CNPhoneNumber>) -> String in
+                return ob.value.stringValue
+            }
         }
 
-        return [
+        var emails: [NSDictionary] = [];
+        if(options.emails) {
+            emails = self.getEmailAddresses();
+        }
+
+        var result: [String : Any] = [
             "id": self.contact.identifier,
-            "firstName": self.contact.givenName,
-            "middleName": self.contact.middleName,
-            "familyName": self.contact.familyName,
-            "phoneNumbers": phoneNumbers
+            "phoneNumbers": phoneNumbers,
+            "emails": emails
         ];
+
+        if(options.firstName) {
+            result["firstName"] = self.contact.givenName;
+        }
+        if(options.middleName) {
+            result["middleName"] = self.contact.middleName;
+        }
+        if(options.familyName) {
+            result["familyName"] = self.contact.familyName;
+        }
+
+        return result as NSDictionary;
     }
 }