+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
 
+## 1.1.0
+- Fields can be specified (**Breaking**: phoneNumbers disabled by default)
+- emails field added 
+
 ## 1.0.0
 - Initial Release